Flyttade Allorders structen till en egen fil så den går att återanvända,
flyttade Lyssnare till firestoremanager så att alla kan använda den
