+++ ResursKollen/ResursKollenApp.swift
@@ -20,7 +20,7 @@ struct ResursKollenApp: App {
   var body: some Scene {
     WindowGroup {
       NavigationView {
-          EmployeeHomeView()
+          AllOrdersView()
       }
     }
   }
