+++ ResursKollen/Services/FirestoreManager.swift
@@ -26,23 +26,24 @@ class FirestoreManager {
         )
     }
     
-    func listenToOrderCollection() {
-        
-        orderRef.addSnapshotListener {snapshot, error in
-            if let error = error {
-                print("Error listening to Orders: \(error.localizedDescription)")
-                return
+    //Snapshot lyssnare för order collectionen som kallas i viewmodels och använder closure i viewmodel
+    func listenToOrderCollection(onUpdate: @escaping ([Order]) -> Void) {
+            orderRef.addSnapshotListener { snapshot, error in
+                if let error = error {
+                    print("Error listening to Orders: \(error.localizedDescription)")
+                    return
+                }
+
+                guard let documents = snapshot?.documents else {
+                    print("No documents")
+                    return
+                }
+
+                let orders = documents.compactMap { doc in
+                    try? doc.data(as: Order.self)
+                }
+
+                onUpdate(orders)
             }
-            
-            guard let documents = snapshot?.documents else {
-                print("No documents")
-                return
-            }
-            
-            self.orders = documents.compactMap { doc in
-                try? doc.data(as: Order.self)
-            }
-            
-        }
     }
 }
