--- ResursKollen/View/ManagerAnnouncmentView.swift
@@ -1,18 +1,0 @@
-//
-//  ManagerAnnouncmentView.swift
-//  ResursKollen
-//
-//  Created by Robin jakobsson on 2025-05-19.
-//
-
-import SwiftUI
-
-struct ManagerAnnouncmentView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    ManagerAnnouncmentView()
-}
